+++ packages/terra-date-time-picker/tests/wdio/date-time-picker-spec.js
@@ -299,37 +299,4 @@ describe('DateTimePicker', () => {
 
     Terra.should.matchScreenshot({ viewports });
   });
-
-  describe('Clears the default date and time on date input focus when default date is out of range', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-default-date-out-of-range');
-      // Removes the blinking cursor to prevent screenshot mismatches.
-      browser.execute('document.querySelector(\'input[name="terra-date-input"]\').style.caretColor = "transparent";');
-      browser.click('input[name="terra-date-input"]');
-    });
-
-    Terra.should.matchScreenshot({ viewports });
-  });
-
-  describe('Clears the default date and time on hour input focus when default date is out of range', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-default-date-out-of-range');
-      // Removes the blinking cursor to prevent screenshot mismatches.
-      browser.execute('document.querySelector(\'input[name="terra-time-hour-input"]\').style.caretColor = "transparent";');
-      browser.click('input[name="terra-time-hour-input"]');
-    });
-
-    Terra.should.matchScreenshot({ viewports });
-  });
-
-  describe('Clears the default date and time on minute input focus when default date is out of range', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-default-date-out-of-range');
-      // Removes the blinking cursor to prevent screenshot mismatches.
-      browser.execute('document.querySelector(\'input[name="terra-time-minute-input"]\').style.caretColor = "transparent";');
-      browser.click('input[name="terra-time-minute-input"]');
-    });
-
-    Terra.should.matchScreenshot({ viewports });
-  });
 });